requirements.txt
torch  
transformers  
Pillow  
numpy  
opencv-python  
matplotlib  
pandas  
